Refactor device routes into dedicated service layer
